refactor function, performance improvements
